restructured code, adjusted imports and setup.py for correct packaging
