+++ src/bjointsp/fixed/fixed_instance.py
