+++ src/bjointsp/network/links.py
