+++ src/bjointsp/network/nodes.py
