+++ src/bjointsp/overlay/edge.py
