+++ src/bjointsp/parameters/networks/square_network.csv
