+++ src/bjointsp/parameters/optional/events.csv
