+++ src/bjointsp/parameters/optional/fixed.csv
