+++ src/bjointsp/parameters/scenarios/mip1-scenario.csv
