+++ src/bjointsp/parameters/scenarios/mip2-scenario.csv
