+++ src/bjointsp/parameters/scenarios/mip3-scenario.csv
