+++ src/bjointsp/parameters/scenarios/simple.csv
