+++ src/bjointsp/parameters/templates/cdn1.csv
