+++ src/bjointsp/parameters/templates/cdn2.csv
