+++ src/bjointsp/template/arc.py
