+++ src/bjointsp/template/component.py
